pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-task-runner"
version = "0.1.0"
description = "Automated task runner that fetches tasks from GitHub Issues and executes them via Claude Code"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "claude-agent-sdk>=0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/task_runner"]

[project.scripts]
task-runner = "task_runner.__main__:main"
